Moves duplicated logic to the base analysis view. Fixes event propagation
