+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/base-layer-analysis-view.js
@@ -1,4 +1,5 @@
 var cdb = require('cartodb-deep-insights.js');
+var LayerAnalysisDraggableHelperView = require('../layer-analysis-draggable-helper-view');
 
 /**
  *  Base layer analysis view.
@@ -7,15 +8,24 @@ module.exports = cdb.core.View.extend({
 
   tagName: 'li',
 
-  events: {
-    'click': '_onNodeClicked'
-  },
-
   className: 'js-analysis',
 
-  _onNodeClicked: function (e) {
-    e.stopPropagation();
+  _onClick: function (e) {
     this.trigger('nodeClicked', this.model.id, this);
+  },
+
+  _addDraggableHelper: function () {
+    this.draggableHelperView = new LayerAnalysisDraggableHelperView({
+      el: this.el
+    });
+
+    this.addView(this.draggableHelperView);
+    this.draggableHelperView.bind('dropped', this._onDropped, this);
+    this.draggableHelperView.bind('click', this._onClick, this);
+  },
+
+  _onDropped: function () {
+    console.log('TODO: creates a new layer from ' + this.model.toJSON()); // TODO: replace with actual layer generation
   }
 
 });
