+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/composite-layer-analysis-view.js
@@ -1,5 +1,4 @@
 var template = require('./composite-layer-analysis-view.tpl');
-var LayerAnalysisDraggableHelperView = require('../layer-analysis-draggable-helper-view');
 var DefaultLayerAnalysisView = require('../analysis-views/default-layer-analysis-view');
 var RefLayerAnalysisView = require('../analysis-views/ref-layer-analysis-view');
 var BaseLayerAnalysisView = require('./base-layer-analysis-view');
@@ -17,6 +16,7 @@ module.exports = BaseLayerAnalysisView.extend({
   tagName: 'li',
 
   className: 'js-analysis Editor-ListAnalysis-item',
+
   initialize: function (opts) {
     if (!opts.layerDefinitionModel) throw new Error('layerDefinitionModel is required');
     if (!opts.analysisDefinitionNodesCollection) throw new Error('analysisDefinitionNodesCollection is required');
@@ -40,15 +40,6 @@ module.exports = BaseLayerAnalysisView.extend({
     }
 
     return this;
-  },
-
-  _addDraggableHelper: function () {
-    this.draggableHelperView = new LayerAnalysisDraggableHelperView({
-      el: this.el
-    });
-
-    this.addView(this.draggableHelperView);
-    this.draggableHelperView.bind('dropped', this._onDropped, this);
   },
 
   _addPrimarySourceView: function () {
@@ -80,10 +71,6 @@ module.exports = BaseLayerAnalysisView.extend({
 
     this.addView(view);
     this.$('.js-secondary-source').append(view.render().el);
-  },
-
-  _onDropped: function () {
-    console.log('TODO: creates a new layer from ' + this.model.toJSON()); // TODO: replace with actual layer generation
   }
 
 });
