+++ lib/assets/javascripts/cartodb3/editor/layers/analysis-views/default-layer-analysis-view.js
@@ -1,5 +1,4 @@
 var template = require('./default-layer-analysis-view.tpl');
-var LayerAnalysisDraggableHelperView = require('../layer-analysis-draggable-helper-view');
 var BaseLayerAnalysisView = require('./base-layer-analysis-view');
 
 /**
@@ -34,19 +33,6 @@ module.exports = BaseLayerAnalysisView.extend({
     }
 
     return this;
-  },
-
-  _addDraggableHelper: function () {
-    this.draggableHelperView = new LayerAnalysisDraggableHelperView({
-      el: this.el
-    });
-
-    this.addView(this.draggableHelperView);
-    this.draggableHelperView.bind('dropped', this._onDropped, this);
-  },
-
-  _onDropped: function () {
-    console.log('TODO: creates a new layer from ' + this.model.toJSON()); // TODO: replace with actual layer generation
   }
 
 });
