+++ lib/assets/javascripts/cartodb3/editor/layers/layer-view.js
@@ -18,7 +18,7 @@ module.exports = cdb.core.View.extend({
 
   events: {
     'click .js-remove': '_onRemove',
-    'click': '_onEditLayer'
+    'click .js-title': '_onEditLayer'
   },
 
   initialize: function (opts) {
